Modifications

+++ src/main/java/org/dcu/cloud/assignment/tfidf/Utility.java
@@ -104,8 +104,8 @@ public class Utility {
     }
 
     private static double calculateTFIDFForCurrentWordWrtDocument(int wordCountInThisPost,int totalWordsInThisPost,int frequencyOfThisWordAcrossWholePostsByUser,int totalPostByUser){
-        double termFrequency=wordCountInThisPost/totalWordsInThisPost;
-        double inverseDocumentFrequency=Math.log(totalPostByUser/frequencyOfThisWordAcrossWholePostsByUser);
+        double termFrequency=(double) wordCountInThisPost/(double) totalWordsInThisPost;
+        double inverseDocumentFrequency=Math.log((double) totalPostByUser/(double) frequencyOfThisWordAcrossWholePostsByUser);
         double tfIdfForThisWordInThisDocument=termFrequency*inverseDocumentFrequency;
         return tfIdfForThisWordInThisDocument;
     }
